conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py36_0
ansimarkup=1.4.0=pypi_0
astor=0.8.0=py36_0
beautifulsoup4=4.8.2=py36_0
better-exceptions-fork=0.2.1.post6=pypi_0
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2019.11.28=py36_0
chardet=3.0.4=pypi_0
click=7.0=pypi_0
colorama=0.4.3=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=pypi_0
decorator=4.4.1=pypi_0
dh-segment-rsl-fork=0.6.1=pypi_0
docopt=0.6.2=pypi_0
gast=0.3.3=py_0
geos=3.8.0=h33f27b4_0
grpcio=1.27.2=py36h351948d_0
h5py=2.10.0=py36h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.9=pypi_0
imageio=2.8.0=pypi_0
intel-openmp=2020.0=166
joblib=0.14.1=pypi_0
jsonpickle=0.9.6=pypi_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=pypi_0
libiconv=1.15=vc14h29686d3_5
libprotobuf=3.11.4=h7bd577a_0
libxml2=2.9.9=h464c3ec_0
libxslt=1.1.33=h579f668_0
lxml=4.5.0=py36h1350720_0
markdown=3.1.1=py36_0
matplotlib=3.1.3=pypi_0
mkl=2020.0=166
mkl-service=2.3.0=py36hb782905_0
mkl_fft=1.0.15=py36h14836fe_0
mkl_random=1.1.0=py36h675688f_0
mock=4.0.1=py_0
munch=2.5.0=pypi_0
networkx=2.4=pypi_0
numpy=1.18.1=pypi_0
numpy-base=1.18.1=py36hc3f5095_1
opencv-python=4.2.0.32=pypi_0
openssl=1.1.1=he774522_0
pandas=1.0.1=pypi_0
pillow=7.0.0=pypi_0
pip=20.0.2=py36_1
protobuf=3.11.4=py36h33f27b4_0
py-cpuinfo=5.0.0=pypi_0
pygments=2.5.2=pypi_0
pyparsing=2.4.6=pypi_0
pyreadline=2.1=py36_1
python=3.6.10=h9f7ef89_0
python-dateutil=2.8.1=pypi_0
pytz=2019.3=pypi_0
pywavelets=1.1.1=pypi_0
requests=2.23.0=pypi_0
sacred=0.7.4=pypi_0
scikit-image=0.16.2=pypi_0
scikit-learn=0.22.1=pypi_0
scipy=1.4.1=py36h9439919_0
setuptools=45.2.0=py36_0
shapely=1.7.0=pypi_0
six=1.14.0=py36_0
soupsieve=1.9.5=py36_0
sqlite=3.31.1=he774522_0
tensorboard=1.13.1=py36_0
tensorflow=1.13.1=gpu_py36h9006a92_0
tensorflow-base=1.13.1=gpu_py36h871c8ca_0
tensorflow-estimator=1.13.0=py_0
tensorflow-gpu=1.13.1=h0d30ee6_0
termcolor=1.1.0=py36_1
tqdm=4.43.0=pypi_0
urllib3=1.25.8=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
werkzeug=1.0.0=py_0
wheel=0.34.2=py36_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.12.0=pypi_0
zlib=1.2.11=h62dcd97_3
